Normalize data model
Instead of keeping match number and type in Event, Match keeps it.  This does
open up the possibility of data inconsistency with regard to match numbers, but
the data changes infrequently and the benefit far outweighs the cost.

Also, the team number is the "_id" of teams now.  The decision to make it a
separate field was born out of my own laziness.  With the Go rewrite, I will
correct it.